Migrated hardware disk size to double

+++ core/src/main/java/de/uniulm/omi/cloudiator/sword/domain/HardwareFlavor.java
@@ -47,5 +47,5 @@ public interface HardwareFlavor extends Resource {
    *
    * @return the amount of disk space or null.
    */
-  Optional<Float> gbDisk();
+  Optional<Double> gbDisk();
 }
